[230922] 전체 타입 다시 복귀

+++ src/main/java/com/winterfoodies/winterfoodies_project/dto/cart/CartDto.java
@@ -12,17 +12,17 @@ import java.util.ArrayList;
 @Setter
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class CartDto {
-    private double id;
+    private Long id;
 
     private String name;
 
     private String storeName;
 
-    private double price;
+    private Long price;
 
-    private double quantity;
+    private Long quantity;
 
-    private double totalPrice;
+    private Long totalPrice;
 
     private String EstimatedCookingTime;
