+++ src/main/java/com/winterfoodies/winterfoodies_project/repository/StoreRepository.java
@@ -22,7 +22,7 @@ public interface StoreRepository extends JpaRepository<Store, Long> {
     @Query("SELECT s FROM Store s JOIN s.storeDetail sd JOIN s.storeProducts sp WHERE " +
             "sp.product.id = :productId AND " + // productId에 해당하는 것만 가져오도록
             "6371 * acos(cos(radians(:latitude)) * cos(radians(sd.latitude)) * cos(radians(sd.longitude) - radians(:longitude)) + sin(radians(:latitude)) * sin(radians(sd.latitude))) < :radius")
-    List<Store> findNearbyStoresByProductId(@Param("productId") double productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
+    List<Store> findNearbyStoresByProductId(@Param("productId") Long productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
 
 
 //    // 판매량 순으로 정렬하여 가게목록 가져오기 (신천3주문 - 소새울2주문 - 대야1주문)
@@ -37,41 +37,41 @@ public interface StoreRepository extends JpaRepository<Store, Long> {
 //    @Query("SELECT s FROM Store s JOIN s.orders o JOIN s.storeProducts sp WHERE "+
 //            "sp.product.id = :productId " + // productId에 해당하는 것만 가져오도록
 //            "GROUP BY s.id")
-//    List<Store> getStoresSortedByMenuSalesByProductId(@Param("productId") double productId);
+//    List<Store> getStoresSortedByMenuSalesByProductId(@Param("productId") Long productId);
 
     @Query("SELECT s FROM Store s JOIN s.storeDetail sd JOIN s.storeProducts sp WHERE " +
             "sp.product.id = :productId AND " + // productId에 해당하는 것만 가져오도록
             "6371 * acos(cos(radians(:latitude)) * cos(radians(sd.latitude)) * cos(radians(sd.longitude) - radians(:longitude)) + sin(radians(:latitude)) * sin(radians(sd.latitude))) < :radius " +
             "GROUP BY s.id")
-    List<Store> getStoresSortedByMenuSalesByProductId(@Param("productId") double productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
+    List<Store> getStoresSortedByMenuSalesByProductId(@Param("productId") Long productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
 
 
     // 4. 상품별, 리뷰순으로 정렬하여 가게목록 가져오기  - 쿼리 삭제하고 프론트에서 처리하기로 함 8/26
 //    @Query("SELECT s FROM Store s JOIN s.reviews r JOIN s.storeProducts sp WHERE "+
 //            "sp.product.id = :productId  " + // productId에 해당하는 것만 가져오도록
 //            "GROUP BY s.id")
-//    List<Store> getStoreByReviewsByProductId(@Param("productId") double productId);
+//    List<Store> getStoreByReviewsByProductId(@Param("productId") Long productId);
 
     @Query("SELECT s FROM Store s JOIN s.storeDetail sd JOIN s.storeProducts sp JOIN s.reviews r WHERE " +
             "sp.product.id = :productId AND " +
             "6371 * acos(cos(radians(:latitude)) * cos(radians(sd.latitude)) * cos(radians(sd.longitude) - radians(:longitude)) + sin(radians(:latitude)) * sin(radians(sd.latitude))) < :radius " +
             "GROUP BY s.id " +
             "ORDER BY COUNT(r.id) DESC")
-    List<Store> getStoreByReviewsByProductId(@Param("productId") double productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
+    List<Store> getStoreByReviewsByProductId(@Param("productId") Long productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
 
 
     // 5. 별점순으로 가게목록 가져오기 - 230830 추가
 //    @Query("SELECT s FROM Store s JOIN s.storeDetail sd JOIN s.storeProducts sp WHERE "+
 //            "sp.product.id = :productId  " + // productId에 해당하는 것만 가져오도록
 //            "ORDER BY sd.averageRating")
-//    List<Store> getStoreByAverageRatingByProductId(@Param("productId") double productId);
+//    List<Store> getStoreByAverageRatingByProductId(@Param("productId") Long productId);
 
 
     @Query("SELECT s FROM Store s JOIN s.storeDetail sd JOIN s.storeProducts sp WHERE " +
             "sp.product.id = :productId AND " +
             "6371 * acos(cos(radians(:latitude)) * cos(radians(sd.latitude)) * cos(radians(sd.longitude) - radians(:longitude)) + sin(radians(:latitude)) * sin(radians(sd.latitude))) < :radius " +
             "ORDER BY sd.averageRating DESC")
-    List<Store> getStoreByAverageRatingByProductId(@Param("productId") double productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
+    List<Store> getStoreByAverageRatingByProductId(@Param("productId") Long productId, @Param("latitude") double latitude, @Param("longitude") double longitude, @Param("radius") double radius);
 
 
 
@@ -90,5 +90,5 @@ public interface StoreRepository extends JpaRepository<Store, Long> {
     @Query("SELECT s FROM Store s WHERE s.storeDetail.addressNo = :addressNo")
     List<Store> searchStoresByAddressNo(@Param("addressNo") String addressNo);
 
-    Store getStoreById(double id);
+    Store getStoreById(Long id);
 }
